+++ eSwitch/main.swift
@@ -48,10 +48,6 @@ class AppDelegate: NSObject, NSApplicationDelegate {
 }
 
 // MARK: - Main
-if CommandLine.arguments.contains("--selftest-ctrl") {
-    runCtrlSelfTest()
-    exit(0)
-}
 if CommandLine.arguments.contains("--selftest-index") {
     runIndexSelfTest()
     exit(0)